mysite/requirements.txt
asgiref==3.4.1
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.1.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
